requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#


-i https://pypi.org/simple/
aiofiles==0.8.0
aniso8601==9.0.1
anyio==3.6.1; python_full_version >= '3.6.2'
asgiref==3.5.2; python_version >= '3.7'
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.2
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
cryptography==37.0.4
dnspython==2.2.1; python_version >= '3.6' and python_version < '4'
ecdsa==0.18.0b2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.2.1
fastapi[all]==0.58.1
graphene==3.1
graphql-core==3.2.1; python_version >= '3.6' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
greenlet==2.0.0a2; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.13.0; python_version >= '3.6'
httptools==0.2.0
idna==3.3; python_version >= '3.5'
itsdangerous==2.1.2
jinja2==3.1.0
markupsafe==2.1.1; python_version >= '3.7'
orjson==3.6.7
passlib[bcrypt]==1.7.4
peewee==3.14.10
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic[dotenv,email]==1.9.0
python-dotenv==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyyaml==6.0
requests==2.28.1
rsa==4.8; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sqlalchemy2-stubs==0.0.2a24; python_version >= '3.6'
sqlalchemy==1.4.39; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlmodel==0.0.6
starlette==0.13.4; python_version >= '3.6'
typing-extensions==4.3.0; python_version >= '3.7'
ujson==5.4.0
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.17.5
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3